Rez Synth: reimplement "legato" mode by giving legato its own voice instance, separate from the array of voices associated with MIDI notes (fixes #8)

+++ dfx-library/dfxsettings.cpp
@@ -923,7 +923,7 @@ void DfxSettings::handleAllNotesOff(int inMidiChannel, unsigned long inOffsetFra
 		return;
 	}
 
-	for (long i = 0; i < DfxMidi::kNumNotes; i++)
+	for (int i = 0; i < DfxMidi::kNumNotes; i++)
 	{
 		handleMidi_automateParams(dfx::MidiEventType::Note, inMidiChannel, i, 0, inOffsetFrames, true);
 	}
